refactor setCell.ts
